fix mocha test
