+++ config/index.js
@@ -5,7 +5,7 @@ dotenv.config();
 
 const config = {
   port: process.env.PORT,
-  dbURL: process.env.dbURL,
+  dbURL: process.env.DB_URL,
 };
 
 mongoose.connect(config.dbURL);
