+++ package.json
@@ -4,10 +4,11 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "ava tests/**/*.js --verbose",
+    "test": "mocha --require babel-register tests/**/*.js --timeout 10000 --exit",
     "test:watch": "ava tests/**/*.js --verbose --watch",
     "ava": "ava --verbose",
-    "dev": "pm2 start app.js --watch --interpreter babel-node"
+    "dev": "pm2 start app.js --watch --interpreter babel-node",
+    "start-dev": "babel-node app.js"
   },
   "repository": {
     "type": "git",
@@ -24,6 +25,7 @@
     "babel-cli": "^6.26.0",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.7.0",
+    "bcryptjs": "^2.4.3",
     "celebrate": "^7.0.6",
     "dotenv": "^5.0.1",
     "express": "^4.16.3",
@@ -31,11 +33,13 @@
     "mongoose": "^5.1.1"
   },
   "devDependencies": {
-    "ava": "^0.25.0",
+    "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
     "eslint": "^4.19.1",
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-plugin-import": "^2.11.0",
     "faker": "^4.1.0",
+    "mocha": "^5.1.1",
     "pm2": "^2.10.3"
   }
 }
